requirements.txt
confluent_kafka==2.3.0
fastavro==1.9.3
google-search-results==2.4.2
langchain_core==0.1.15
langchain_openai==0.0.3
langchain-community==0.0.15
langchain==0.1.3
openai==1.9.0
psycopg2-binary==2.9.9
requests==2.31.0